Documentation de la classe date.

+++ kernel/framework/util/date.class.php
@@ -47,7 +47,7 @@ define('TIMEZONE_AUTO', TIMEZONE_USER);
  * @desc This class allows you to handle easily some dates. A date is a day and an hour (year, month, day, hour, minutes, seconds).
  * It supports the most common formats and manages timezones. Here are the definitions of the 3 existing timezones:
  * <ul>
- * 	<li>Server timezone: it's the timezone of the server, configured by the hoster. For instance, if your server is in France, it should be GMT+1.</li>
+ * 	<li>System timezone: it's the timezone of the server, configured by the hoster. For instance, if your server is in France, it should be GMT+1.</li>
  * 	<li>Site timezone: it's the timezone of the central place of the site. For example, if your site deals with the italian soccer championship, it will be GMT+1.</li>
  * 	<li>User timezone :  each registered user can specify its timezone. It's particulary useful for people who visit some sites from a foreign country.</li>
  * @author Benoit Sautel <ben.popeye@phpboost.com>
@@ -56,13 +56,43 @@ class Date
 {
 	/**
 	 * @desc Builds and initializes a date. It admits a variable number of parameters depending on the value of the first one.
-	 * The first parameter determines how to initialize the date:
+	 * The second parameter allows us to chose what time referential we use to create the date:
 	 * <ul>
-	 * 	<li>DATE_NOW will initialize the date to the current time</li>
-	 * 	<li>DATE_YEAR_MONTH_DAY
+	 * 	<li>TIMEZONE_SYSTEM if that date comes from for example the database (dates must be stored under this referential).</li>
+	 * 	<li>TIMEZONE_SITE if it's an entry coming from the site (nearly never used).</li>
+	 * 	<li>TIMEZONE_USER if it's an entry coming from the user (it's own timezone will be used)</li>
+	 * </ul>
+	 * The first parameter determines how to initialize the date, here are the rules to use for the other parameters:
+	 * <ul>
+	 * 	<li>DATE_NOW will initialize the date to the current time.
+	 * $date = new Date(DATE_NOW); will build a date with the current date.</li>
+	 * 	<li>DATE_YEAR_MONTH_DAY if you want to build a date from a specified day (year, month, day). In this case the following parameters are:
+	 * 		<ul>
+	 * 			<li>int The year (for instance 2009)</li>
+	 * 			<li>int The month (for example 11)</li>
+	 * 			<li>int The day (for example 09)</li>
+	 * 		</ul>
+	 * For example, $date = new Date(DATE_YEAR_MONTH_DAY, TIMEZONE_USER, 2009, 11, 09);
+	 * 	</li>
+	 * 	<li>DATE_YEAR_MONTH_DAY_HOUR_MINUTE_SECOND if you want to build a date from a specified time. Here is the rule for the following parameters:
+	 * 		<ul>
+	 * 			<li>int The year (for instance 2009)</li>
+	 * 			<li>int The month (for instance 11)</li>
+	 * 			<li>int The day (for instance 09)</li>
+	 * 			<li>int The hour (for instance 12)</li>
+	 * 			<li>int The minutes (for instance 34)</li>
+	 * 			<li>int The seconds (for instance 12)</li>
+	 * 		</ul>
+	 * For instance $date = new Date(DATE_YEAR_MONTH_DAY_HOUR_MINUTE_SECOND, 2009, 11, 09, 12, 34, 12);
+	 * 	</li>
+	 * 	<li>DATE_TIMESTAMP which builds a date from a UNIX timestamp.
+	 * For example $date = new Date(DATE_TIMESTAMP, time()); is equivalent to $date = new Date(DATE_NOW);</li>
+	 * 	<li>DATE_FROM_STRING which decodes a date written in a string by matching a pattern you have to specify.
+	 * The pattern is easy to write: d for day, m for month and y for year.
+	 * For instance, if your third parameter is '24/12/2009' and the fourth is 'm/d/y', it will be the december 24th of 2009.</li>
+	 * </ul>
 	 * Here are the rules:
 	 * 
-	 * @return unknown_type
 	 */
 	function Date()
 	{
@@ -158,7 +188,25 @@ class Date
 		}
 	}
 	
-	// Fonction qui retourne la date formatée
+	/**
+	 * @desc Formats the date to a particular format.
+	 * @param int $format One of the following enumeration:
+	 * <ul>
+	 * 	<li>DATE_FORMAT_TINY for a tiny formatting (only month and day)</li>
+	 * 	<li>DATE_FORMAT_SHORT for a short formatting (month, day, year)</li>
+	 * 	<li>DATE_FORMAT for a longer displaying (year, month, day, hour and minutes)</li>
+	 * 	<li>DATE_FORMAT_LONG for a total displaying (year, month, day, hour, minutes and seconds)</li>
+	 * 	<li>DATE_RFC822_F to format according to what the RFC822 announces</li>
+	 * 	<li>DATE_RFC3339_F to format according to what the RFC3339 announces</li>
+	 * </ul> 
+	 * @param int $referencial_timezone One of the following enumeration:
+	 * <ul>
+	 * 	<li>TIMEZONE_SYSTEM</li>
+	 * 	<li>TIMZONE_SITE</li>
+	 * 	<li>TIMEZONE_USER</li>
+	 * </ul>
+	 * @return string The formatted date
+	 */
 	function format($format = DATE_FORMAT_TINY, $referencial_timezone = TIMEZONE_USER)
 	{
 		global $LANG, $CONFIG;
@@ -202,64 +250,98 @@ class Date
 		}
 	}
 	
-	//Renvoie le timestamp
+	/**
+	 * @desc Returns the timestamp associated to the date
+	 * @return int The timestamp
+	 */
 	function get_timestamp()
 	{
 		return $this->timestamp;
 	}
 	
-	// Renvoie l'année
+	/**
+	 * @desc Returns the year of the date
+	 * @return int The year
+	 */
 	function get_year()
 	{
-		return date('Y', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
+		return (int)date('Y', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
 	}
 
-	// Renvoie le mois
+	/**
+	 * @desc Returns the month of the date
+	 * @return int The month
+	 */
 	function get_month()
 	{
-		return date('m', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
+		return (int)date('m', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
 	}
 	
-	//Renvoie le jour
+	/**
+	 * @desc Returns the day of the date
+	 * @return int The day
+	 */
 	function get_day()
 	{
-		return date('d', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
+		return (int)date('d', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
 	}
 	
-	//Renvoie l'heure
+	/**
+	 * @desc Returns the hours of the date
+	 * @return int The hours
+	 */
 	function get_hours()
 	{
-		return date('H', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
+		return (int)date('H', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
 	}
 	
-	//Renvoie les minutes
+	/**
+	 * @desc Returns the minutes of the date
+	 * @return int The minutes
+	 */
 	function get_minutes()
 	{
-		return date('i', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
+		return (int)date('i', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
 	}
 	
-	//Renvoie les secondes
+	/**
+	 * @desc Returns the seconds of the date
+	 * @return int The seconds
+	 */
 	function get_seconds()
 	{
-		return date('s', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
+		return (int)date('s', $this->timestamp + $this->_compute_server_user_difference(TIMEZONE_USER) * 3600);
 	}
 	
-	//Renvoie une chaine au format YYYY-mm-dd
+	/**
+	 * @desc Exports the date according to the format YYYY-mm-dd
+	 * @return string The formatted date
+	 */
 	function to_date()
 	{
 		return date('Y-m-d', $this->timestamp);
 	}
 	
 	# This should be static#
-	//Function which determines wether a date is correct
+	/**
+	 * @static
+	 * @desc Determines wether a date is correct. For example the february 31st is not correct.
+	 * @param int $month The month
+	 * @param int $day The day
+	 * @param int $year The year
+	 * @return bool true if the date is correct and false otherwise.
+	 */
 	function check_date($month, $day, $year)
 	{
 		return checkdate($month, $day, $year);
 	}
 	
 	## Private ##
-	//Renvoie le nombre d'heures de décalage par rapport à un certain référentiel
-	#This should be static#
+	/**
+	 * @static
+	 * @desc Computes the time difference between the server and the current user
+	 * @return int The time difference (in hours)
+	 */
 	function _compute_server_user_difference($referencial_timezone = 0)
 	{
 		global $CONFIG, $User;
@@ -289,7 +371,9 @@ class Date
 	    return $timezone;
 	}
 	
-	// Timestamp correspondant à la date
+	/**
+	 * @var int The timestamp of the current date
+	 */
 	var $timestamp = 0;
 }
